+++ apps/backend-ops/src/main.ts
@@ -9,7 +9,7 @@ import {
   storeSearchableMovie
 } from '../../backend-functions/src/internals/algolia';
 import { firebase } from '@env';
-import { updateInvitationDocument, updateOrganizationDocument, updatePicturesMovieDocument, updateAvatarUserDocument } from './firestoreMigrations/ToV2';
+import { updateInvitationDocument, updateOrganizationDocument, updatePicturesMovieDocument } from './firestoreMigrations/0002';
 
 async function prepareForTesting() {
   console.info('Syncing users...');
@@ -66,12 +66,11 @@ async function upgradeAlgoliaMovies() {
   return Promise.all(promises);
 }
 
-function upgradeToV2() {
+function migrateToV2() {
   const { db } = loadAdminServices();
   updateInvitationDocument(db);
   updateOrganizationDocument(db);
   updatePicturesMovieDocument(db);
-  updateAvatarUserDocument(db);
 }
 
 const args = process.argv.slice(2);
@@ -89,6 +88,6 @@ if (cmd === 'prepareForTesting') {
   syncUsers(USERS).then(() => process.exit(0));
 } else if (cmd === 'upgradeAlgoliaMovies') {
   upgradeAlgoliaMovies();
-} else if (cmd === 'prepareToV2') {
-  prepareToV2();
+} else if (cmd === 'migrateToV2') {
+  migrateToV2();
 }
